Add Text Field to Item
Added a multiline text field to Items.
This field is managed, so the data can accessed from outside the object which
may be necessary for the drag and drop functionality.
The field component is from material design but has been styled for aesthetics.
Changed css class name to better conform with BEM.
Made remove item button background transparent so it would not overlap text.
Made remove item button always on top as text could prevent it from being
clicked previously.