Add Colors To Items
Added a color picker to item cards.
Implements custom color picker to facilitate this feature.
Added color to item state, so that it can be kept track of when dragging between lists.
This color is saved as an index of the colors array which is defined in the ColorPicker
component. This is so that colors are all defined in one place, and Items can be created
with a default index of 0, without needing to change everytime colors are changed.